feat: Add code for fetching product sales data from BigQuery
This commit adds a new file, main.py, to the product_sales/fetch directory. The main.py file contains code for fetching product sales data from BigQuery using a Cloud Function. The code initializes a BigQuery client, specifies the table to query, and executes the query to retrieve the data. The fetched data is then converted to JSON format and returned as a response.

The purpose of this code change is to enable the retrieval of product sales data from BigQuery for further processing or analysis.