refactor: Update CD workflow to include all branches
Simplify the CD workflow by removing the unnecessary "run-name" field and updating the "on" section to include all branches (main and develop). This ensures that the CD workflow is triggered whenever a CI workflow is completed on any of these branches. The environment inputs and event triggers have also been updated accordingly.

Closes #3